backend/requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-ollama
langchain-community
python-dotenv
pydantic
websockets
